Step 6 - moved useEscapeKey to custom hook
