+++ src/components/ToastPlayground/ToastPlayground.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import Button from '../Button';
-import Toast from '../Toast';
 import styles from './ToastPlayground.module.css';
 import { useEffect } from 'react';
 import ToastShelf from '../ToastShelf/ToastShelf';
@@ -10,20 +9,6 @@ import { ToastContext } from '../ToastProvider/ToastProvider';
 const VARIANT_OPTIONS = ['notice', 'warning', 'success', 'error'];
 
 function ToastPlayground() {
-  const [currentVariant, setCurrentVariant] =
-    React.useState('notice');
-
-  const [message, setMessage] = React.useState('');
-
-  const { createdToasts, createToast } = React.useContext(ToastContext);
-
-
-  function submitToast(receivedMessage, receivedVariant) {
-    createToast(receivedMessage, receivedVariant);
-    setMessage('');
-    setCurrentVariant('notice');
-  }
-
   return (
     <div className={styles.wrapper}>
       <header>
@@ -36,69 +21,87 @@ function ToastPlayground() {
 
       <ToastShelf />
 
-      <form
-        className={styles.controlsWrapper}
-        onSubmit={(event) => {
-          event.preventDefault();
-          submitToast(message, currentVariant);
-        }}
-      >
-        <div className={styles.row}>
-          <label
-            htmlFor="message"
-            className={styles.label}
-            style={{ alignSelf: 'baseline' }}
-          >
-            Message
-          </label>
-          <div className={styles.inputWrapper}>
-            <textarea
-              id="message"
-              value={message}
-              className={styles.messageInput}
-              onChange={(event) => {
-                setMessage(event.target.value);
-              }}
-            />
-          </div>
-        </div>
-
-        <div className={styles.row}>
-          <div className={styles.label}>Variant</div>
-          <div
-            className={`${styles.inputWrapper} ${styles.radioWrapper}`}
-          >
-            {VARIANT_OPTIONS.map((option) => (
-              <label
-                htmlFor={`variant-${option}`}
-                key={`${option}`}
-              >
-                <input
-                  id={`variant-${option}`}
-                  type="radio"
-                  name="variant"
-                  value={`${option}`}
-                  checked={option === currentVariant}
-                  onChange={(event) => {
-                    setCurrentVariant(event.target.value);
-                  }}
-                />
-                {option}
-              </label>
-            ))}
-          </div>
-        </div>
-
-        <div className={styles.row}>
-          <div className={styles.label} />
-          <div
-            className={`${styles.inputWrapper} ${styles.radioWrapper}`}
-          >
-            <Button>Pop Toast!</Button>
-          </div>
-        </div>
-      </form>
+      <ToastForm />
     </div>
+  );
+}
+
+function ToastForm() {
+  const [currentVariant, setCurrentVariant] =
+    React.useState('notice');
+
+  const [message, setMessage] = React.useState('');
+
+  const { createToast } = React.useContext(ToastContext);
+
+  function submitToast(receivedMessage, receivedVariant) {
+    createToast(receivedMessage, receivedVariant);
+    setMessage('');
+    setCurrentVariant('notice');
+  }
+  return (
+    <form
+      className={styles.controlsWrapper}
+      onSubmit={(event) => {
+        event.preventDefault();
+        submitToast(message, currentVariant);
+      }}
+    >
+      <div className={styles.row}>
+        <label
+          htmlFor="message"
+          className={styles.label}
+          style={{ alignSelf: 'baseline' }}
+        >
+          Message
+        </label>
+        <div className={styles.inputWrapper}>
+          <textarea
+            id="message"
+            value={message}
+            className={styles.messageInput}
+            onChange={(event) => {
+              setMessage(event.target.value);
+            }}
+          />
+        </div>
+      </div>
+
+      <div className={styles.row}>
+        <div className={styles.label}>Variant</div>
+        <div
+          className={`${styles.inputWrapper} ${styles.radioWrapper}`}
+        >
+          {VARIANT_OPTIONS.map((option) => (
+            <label
+              htmlFor={`variant-${option}`}
+              key={`${option}`}
+            >
+              <input
+                id={`variant-${option}`}
+                type="radio"
+                name="variant"
+                value={`${option}`}
+                checked={option === currentVariant}
+                onChange={(event) => {
+                  setCurrentVariant(event.target.value);
+                }}
+              />
+              {option}
+            </label>
+          ))}
+        </div>
+      </div>
+
+      <div className={styles.row}>
+        <div className={styles.label} />
+        <div
+          className={`${styles.inputWrapper} ${styles.radioWrapper}`}
+        >
+          <Button>Pop Toast!</Button>
+        </div>
+      </div>
+    </form>
   );
 }
