+++ src/hooks/useEscapeKey.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+export function useEscapeKey(callback) {
+  React.useEffect(() => {
+    function handleKeyDown(event) {
+      if (event.code === 'Escape') {
+        callback(event);
+      }
+    }
+    window.addEventListener('keydown', handleKeyDown);
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+    };
+  }, [callback]);
+
+  return;
+}
+
+export default useEscapeKey;
